chore: clean up the code, add inline docs
