+++ .projenrc.js
@@ -51,7 +51,7 @@ const deployTask = project.addTask('integ:exec:deploy');
 deployTask.exec(`rm -fr ${appDir}`);
 deployTask.exec(`cdk --app "${tsNode} test/exec.integ.ts" deploy --output ${appDir}`);
 
-const cliCmd = `${tsNode} src/cli/cdk-exec.ts --app ${appDir} exec`;
+const cliCmd = `${tsNode} src/cli/cdk-exec.ts --app ${appDir}`;
 project.addTask('integ:exec:sfn:succeed', {
   exec: `${cliCmd} integ-cdk-exec/StateMachine --input '{"succeed":true}'`,
 });
