+++ test/exec.integ.snapshot/tree.json
@@ -9,7 +9,7 @@
         "path": "Tree",
         "constructInfo": {
           "fqn": "constructs.Construct",
-          "version": "10.0.61"
+          "version": "10.0.63"
         }
       },
       "integ-cdk-exec": {
@@ -222,7 +222,7 @@
             },
             "constructInfo": {
               "fqn": "constructs.Construct",
-              "version": "10.0.61"
+              "version": "10.0.63"
             }
           }
         },
